fix: rename files
delete "temp" and make the names official
